app/requirements.txt
anyio==3.5.0
asgiref==3.5.0
astor==0.8.1
astroid==2.11.7
attrs==22.1.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
coverage==6.5.0
cryptography==37.0.2
dill==0.3.5.1
fastapi==0.75.0
func-timeout==4.3.5
graphviz==0.20
h11==0.13.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
lazy-object-proxy==1.7.1
mccabe==0.7.0
msal==1.18.0
mysql-connector-python==8.0.28
networkx==2.8.8
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.5.0
PyJWT==2.4.0
pylint==2.14.5
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websocket-client==1.3.1
websockets==10.2
wrapt==1.14.1
